+++ lib/dawn/knowledge_base.rb
@@ -245,6 +245,7 @@ require "dawn/kb/cve_2015_3448"
 require "dawn/kb/cve_2015_4020"
 require "dawn/kb/cve_2015_7576"
 require "dawn/kb/cve_2015_7577"
+require "dawn/kb/cve_2015_7578"
 require "dawn/kb/cve_2015_7579"
 
 # CVE - 2016
@@ -535,6 +536,7 @@ module Dawn
           Dawn::Kb::CVE_2015_4020.new,
           Dawn::Kb::CVE_2015_7576.new,
           Dawn::Kb::CVE_2015_7577.new,
+          Dawn::Kb::CVE_2015_7578.new,
           Dawn::Kb::CVE_2015_7579.new,
           Dawn::Kb::CVE_2016_0751.new,
           Dawn::Kb::CVE_2016_0752.new,