+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -1104,4 +1104,9 @@ it "must have test for CVE-2016-0753" do
   sc.should_not   be_nil
   sc.class.should == Dawn::Kb::CVE_2016_0753
 end
+it "must have test for CVE-2015-7578" do
+  sc = kb.find("CVE-2015-7578")
+  sc.should_not   be_nil
+  sc.class.should == Dawn::Kb::CVE_2015_7578
+end
 end