+++ Changelog.md
@@ -27,6 +27,8 @@ _latest update: Thu Jan 28 23:30:47 CET 2016_
   Vulnerability in Action View
 * Issue #181 - Adding a check for CVE-2016:0753: Possible Input Validation
   Circumvention in Active Model
+* Issue #180 - Adding a check for CVE-2015-7578: Possible XSS vulnerability in
+  rails-html-sanitizer
 
 ## Version 1.5.2 - codename: Tow Mater (2015-12-16)